added the analog data files
