+++ data_processing/README.txt
@@ -1,4 +1,5 @@
-1. for the Demographics spreadsheet run make sure all the gait data is in the current directory inside 'data', or otherwise modify your basepath inside the meta.py file, and then run the following command that generates the CSV file:
+1. for the Demographics spreadsheet run make sure all the gait data is in the current directory inside 'data', or otherwise modify your basepath inside the Demographics.py file, and then run the following command that generates the CSV file:
 
-python meta.py
+python Demographics.py
 
+2. for the Markers, EMG and ForcePlate spreadsheets run their corresponding .ipynb files.
